+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/comment/dto/CommentRequest.kt
@@ -1,0 +1,5 @@
+package com.teamsparta.mini5foodfeed.domain.comment.dto
+
+class CommentRequest(val contents: String) {
+
+}
