+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/comment/dto/CommentResponse.kt
@@ -1,0 +1,5 @@
+package com.teamsparta.mini5foodfeed.domain.comment.dto
+
+class CommentResponse(val contents: String) {
+
+}
