+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/comment/repository/CommentRepository.kt
@@ -1,6 +1,6 @@
-package com.teamsparta.mini5foodfeed.domain.feed.comment.repository
+package com.teamsparta.mini5foodfeed.domain.comment.repository
 
-import com.teamsparta.mini5foodfeed.domain.feed.comment.model.Comment
+import com.teamsparta.mini5foodfeed.domain.comment.model.Comment
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
