--- src/main/kotlin/com/teamsparta/mini5foodfeed/domain/feed/comment/dto/CommentRequest.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.mini5foodfeed.domain.feed.comment.dto
-
-class CommentRequest(val contents: String) {
-
-}
