--- src/main/kotlin/com/teamsparta/mini5foodfeed/domain/feed/comment/dto/CommentResponse.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.mini5foodfeed.domain.feed.comment.dto
-
-class CommentResponse(val contents: String) {
-
-}
