--- src/main/kotlin/com/teamsparta/mini5foodfeed/domain/feed/tag/service/TagService.kt
@@ -1,4 +1,0 @@
-package com.teamsparta.mini5foodfeed.domain.feed.tag.service
-
-interface TagService {
-}
