+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/tag/controller/TagController.kt
@@ -1,0 +1,4 @@
+package com.teamsparta.mini5foodfeed.domain.tag.controller
+
+class TagController {
+}
