+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/tag/repository/TagRepository.kt
@@ -1,0 +1,4 @@
+package com.teamsparta.mini5foodfeed.domain.tag.repository
+
+interface TagRepository {
+}
