+++ src/main/kotlin/com/teamsparta/mini5foodfeed/domain/tag/service/TagService.kt
@@ -1,0 +1,4 @@
+package com.teamsparta.mini5foodfeed.domain.tag.service
+
+interface TagService {
+}
